Refactor large usePermitForm hook to smaller hooks
